requirements.txt
nltk>=3.2.4
numpy>=1.13.0
regex>=2017.6.7
tensorflow>=1.2.0
Pillow>=2.0.0
